Breaking - Update the syntax to have the default style option include an icon
Before, you needed to explicitly say "info-icon" to include the icon. With the new syntax, just by specifying "info", the icon will be included and you need to explicitly say "info no-icon" in order to have the styling without an icon.

I did this because I felt that the icon was more of a default behavior and more widely used. I made it somewhat backwards compatible by leaving the "info-icon" syntax as a valid syntax to display with an icon